fix(checkbox): restore support for `labelHelp` and `fieldHelp` props
This reverts commit 2f2f340fcdfca1acb5b72bad6df7299eece3eb3e.
