+++ src/components/numeral-date/numeral-date.mdx
@@ -55,7 +55,7 @@ For more information check our [Validations](../?path=/docs/documentation-valida
 
 The folowing examples use the new validation pattern that is available by setting the `validationRedesignOptIn` flag on the `CarbonProvider` to true.
 
-**Note:** The `labelHelp` tooltip will be rendered as "Hint text".
+**Note:** The `legendHelp` tooltip will be rendered as "Hint text".
 
 <Canvas of={NumeralDateStories.NewValidation} />
 
@@ -91,7 +91,7 @@ The inline label can change to be top aligned at a breakpoint. Enable this by pa
 
 ### With label help
 
-**Note:** The `labelHelp` tooltip will be rendered as hint text if the `validationRedesignOptIn` flag on the `CarbonProvider` is true.
+**Note:** The `legendHelp` tooltip will be rendered as hint text if the `validationRedesignOptIn` flag on the `CarbonProvider` is true.
 
 <Canvas of={NumeralDateStories.WithLabelHelp} />
