+++ src/components/switch/switch.mdx
@@ -93,7 +93,7 @@ You can use the `isOptional` prop to indicate if the field is optional.
 
 ### With labelHelp
 
-**Note:** The `labelHelp` tooltip will be rendered as hint text if the `validationRedesignOptIn` flag on the `CarbonProvider` is true.
+**Note:** The `legendHelp` tooltip will be rendered as hint text if the `validationRedesignOptIn` flag on the `CarbonProvider` is true.
 
 <Canvas of={SwitchStories.WithLabelHelp} name="with labelHelp" />
 
@@ -155,7 +155,7 @@ It is possible to use the `tooltipPosition` to override the default placement of
 
 The folowing examples use the new validation pattern that is available by setting the `validationRedesignOptIn` flag on the `CarbonProvider` to true.
 
-**Note:** The `labelHelp` tooltip will be rendered as "Hint text".
+**Note:** The `legendHelp` tooltip will be rendered as "Hint text".
 
 <Canvas
   of={SwitchStories.NewValidationString}
